+++ WpfApp/ViewModel/DomainListViewModel.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Net.Http.Json;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Input;
 using Api.Models;
 using Newtonsoft.Json;
@@ -15,7 +16,6 @@ namespace WpfApp.ViewModel
     {
         private ObservableCollection<Domain> _domainsList;
 
-        private bool _visibilityCreateMenu;
         private bool _visibilityEditMenu;
         
         private Domain _selectDomain;
@@ -26,12 +26,6 @@ namespace WpfApp.ViewModel
         {
             get { return _domainsList; }
             set {SetProperty(ref _domainsList , value); }
-        }
-
-        public bool VisibilityCreateMenu
-        {
-            get { return _visibilityCreateMenu; }
-            set {SetProperty(ref _visibilityCreateMenu , value); }
         }
 
         public bool VisibilityEditMenu
@@ -49,21 +43,17 @@ namespace WpfApp.ViewModel
         #endregion
         
         public ICommand ToggleAddMenu { get; }
-        public ICommand CreateDomainCommand { get; }
-        
-        
         public ICommand ToggleEditMenu { get; }
+                
         public ICommand SaveDomainCommand { get; }
-        
-        
+            
         public ICommand DeleteDomainCommand { get; }
 
         public DomainListViewModel()
         {
             ToggleAddMenu = new ViewModelCommand<Object>(ExecuteToggleAddMenu);
-            CreateDomainCommand = new ViewModelCommand<Object>(ExecuteCreateDomainCommand);
-            
             ToggleEditMenu = new ViewModelCommand<Domain>(ExecuteToggleEditMenu);
+            
             SaveDomainCommand = new ViewModelCommand<Object>(ExecuteSaveDomainCommand);
             
             DeleteDomainCommand = new ViewModelCommand<Domain>(DeleteDomain);
@@ -82,33 +72,40 @@ namespace WpfApp.ViewModel
         private void ExecuteToggleAddMenu(Object obj)
         {
             SelectDomain = new Domain();
-            VisibilityCreateMenu = !VisibilityCreateMenu;
-        }
-        private async void ExecuteCreateDomainCommand(object obj)
-        {
-            var response = await ModeCommun.client.PostAsJsonAsync("domain", SelectDomain);
-            VisibilityCreateMenu = !VisibilityCreateMenu;
-            GetDomains();
+            VisibilityEditMenu = true;
         }
 
-        
         private void ExecuteToggleEditMenu(Domain obj)
         {
             SelectDomain = obj;
-            VisibilityEditMenu = !VisibilityEditMenu;
+            VisibilityEditMenu = true;
         }
         private async void ExecuteSaveDomainCommand(Object obj)
         {
-            var response = await ModeCommun.client.PutAsJsonAsync("domain/" + SelectDomain.Id, SelectDomain);
-            VisibilityEditMenu = !VisibilityEditMenu;
-            GetDomains();
+            if (SelectDomain.Id == 0)
+            {
+                var response = await ModeCommun.client.PostAsJsonAsync("domain", SelectDomain);                
+                if (response.StatusCode == System.Net.HttpStatusCode.OK)
+                {
+                    GetDomains();
+                    VisibilityEditMenu = false;
+                }
+
+            }
+            else
+            {
+                var response = await ModeCommun.client.PutAsJsonAsync("domain/" + SelectDomain.Id, SelectDomain);
+            }
         }
         
         
         private async void DeleteDomain(Domain obj)
         {
-            var response = await ModeCommun.client.DeleteAsync("domain/" + obj.Id);
-            GetDomains();
+            if (MessageBox.Show("Are you sure you want to delete this domain?", "Warning", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.Yes) == MessageBoxResult.Yes)
+            {                
+                var response = await ModeCommun.client.DeleteAsync("domain/" + obj.Id);
+                DomainsList.Remove(obj);
+            }
         }
     }
 }
